+++ src/Emby.Plugin.RdZurg/Library/ReleaseNames.cs
@@ -93,11 +93,12 @@ public sealed class ReleaseNames
         + "blu-ray|x264|x265|h264|h265|xvid|xvidvd|xxx|www.www|aac";
 
     // The year is the last 19xx/20xx that has a title in front of it and is not part of a longer number or a date.
-    // A year right after punctuation wins over one after a space, so "1990-1994 2024" reads as 1994.
+    // A year right after punctuation wins over one after a space, so "1990-1994 2024" reads as 1994. A resolution
+    // is not a year: "[BDRip 1920x1080 HEVC]" would otherwise file a release as a 1920 film.
     private static readonly Regex[] _years =
     {
-        new(@"^(?<name>.+[^_,.()\[\]\-])[_.()\[\]\-](?<year>(?:19|20)[0-9]{2})(?![0-9]|\W[0-9]{2}\W[0-9]{2})", Options),
-        new(@"^(?<name>.+[^_,.()\[\]\-])[ _.()\[\]\-]+(?<year>(?:19|20)[0-9]{2})(?![0-9]|\W[0-9]{2}\W[0-9]{2})", Options),
+        new(@"^(?<name>.+[^_,.()\[\]\-])[_.()\[\]\-](?<year>(?:19|20)[0-9]{2})(?![0-9]|[xX][0-9]{3,4}(?![0-9])|\W[0-9]{2}\W[0-9]{2})", Options),
+        new(@"^(?<name>.+[^_,.()\[\]\-])[ _.()\[\]\-]+(?<year>(?:19|20)[0-9]{2})(?![0-9]|[xX][0-9]{3,4}(?![0-9])|\W[0-9]{2}\W[0-9]{2})", Options),
     };
 
     // Applied in order and cumulatively, each result trimmed.